organize interface and service class code
