--- src/main/Application/Interface/Repository/IAvatarRepository.cs
@@ -1,10 +1,0 @@
-﻿using ei8.Cortex.Subscriptions.Domain.Model;
-using System.Threading.Tasks;
-
-namespace ei8.Cortex.Subscriptions.Application.Interface.Repository
-{
-    public interface IAvatarRepository
-    {
-        Task<Avatar> GetOrAddAsync(string url);
-    }
-}
