+++ src/main/Domain.Model/Domain.Model/IAvatarRepository.cs
@@ -1,0 +1,9 @@
+﻿using System.Threading.Tasks;
+
+namespace ei8.Cortex.Subscriptions.Domain.Model
+{
+    public interface IAvatarRepository
+    {
+        Task<Avatar> GetOrAddAsync(string url);
+    }
+}
