+++ src/main/Domain.Model/Domain.Model/IBrowserReceiverRepository.cs
@@ -1,9 +1,8 @@
-﻿using ei8.Cortex.Subscriptions.Domain.Model;
-using System;
+﻿using System;
 using System.Collections.Generic;
 using System.Threading.Tasks;
 
-namespace ei8.Cortex.Subscriptions.Application.Interface.Repository
+namespace ei8.Cortex.Subscriptions.Domain.Model
 {
     public interface IBrowserReceiverRepository
     {
