+++ src/main/Domain.Model/Domain.Model/IUserRepository.cs
@@ -1,0 +1,10 @@
+﻿using System;
+using System.Threading.Tasks;
+
+namespace ei8.Cortex.Subscriptions.Domain.Model
+{
+    public interface IUserRepository
+    {
+        Task<User> GetOrAddAsync(Guid id);
+    }
+}
